requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
supabase>=2.9.0
openai>=1.40.0
pypdf2==3.0.1
python-docx==1.1.2
python-multipart>=0.0.9
pydantic>=2.9.0
python-dotenv==1.0.1
httpx>=0.27.0